producer/requirements.txt
kafka-python    